requirements.txt
Django>=1.4.3
djangorestframework>=2.1.15
msgpack-python>=0.2.4
python-dateutil==2.1